Fix distributed module build issues without breaking master branch
Summary of changes:
- Removed gmock/gtest dependencies from distributed matcher implementations
- Simplified matcher factory functions to basic two-parameter versions
- Fixed uninitialized variable bug in result_aggregation.cpp
- Added missing type aliases for ScenarioDefinition and ScenarioResult
- Disabled incomplete coordinator/node/gRPC implementations (marked TODO)
- Disabled tools build that depend on incomplete distributed implementations
- Core library (libwadjet.a) builds cleanly
- Distributed library (libwadjet_distributed.a) builds cleanly

Build configuration tested:
  cmake -DWADJET_ENABLE_DISTRIBUTED=ON -DWADJET_BUILD_EXAMPLES=OFF -DWADJET_BUILD_TESTS=OFF
  make

Changes are isolated to the distributed module - no modifications to master branch code.
The distributed feature remains optional via -DWADJET_ENABLE_DISTRIBUTED flag.

Fixes:
1. Removed gmock header requirements from distributed_matcher.hpp/cpp
   - testing::Matcher<PacketView> overloads caused -Werror=undef cascade from gtest
   - Simplified to basic factory functions without GoogleTest integration

2. Fixed bugs in result_aggregation.cpp
   - total_passed/total_failed were self-initialized (int x = x)
   - Now properly calculated from node results

3. Added type definitions in coordinator.hpp
   - ScenarioDefinition = DistributedScenario
   - ScenarioResult = wadjet::scenario::ScenarioResult

Pending work (marked with TODO comments):
- Complete coordinator.cpp implementation (~800 lines, has compilation errors)
- Complete node.cpp implementation (incomplete virtual methods)
- Complete gRPC service/client implementations (syntax errors)
- Re-enable tools build when above are complete

+++ include/wadjet/distributed/types.hpp
@@ -120,8 +120,8 @@ struct NodeCaptureResult {
             return 0.0;
         }
         int64_t duration_ns = end_time_ns - start_time_ns;
-        double duration_sec = duration_ns / 1e9;
-        return duration_sec > 0 ? packet_count / duration_sec : 0.0;
+        double duration_sec = static_cast<double>(duration_ns) / 1e9;
+        return duration_sec > 0 ? static_cast<double>(packet_count) / duration_sec : 0.0;
     }
 
     /**
@@ -135,7 +135,8 @@ struct NodeCaptureResult {
         if (packet_count + packet_loss_count == 0) {
             return 0.0;
         }
-        return (100.0 * packet_loss_count) / (packet_count + packet_loss_count);
+        return (100.0 * static_cast<double>(packet_loss_count)) /
+               static_cast<double>(packet_count + packet_loss_count);
     }
 };
